The task passed the test. Fixed the writeToFile() method if an empty list is provided. And also improved the sorting of objects in the list due to reflection. And added deletion of a temporary file after each test if it is available.
