+++ expected.csv
@@ -1,9 +1,9 @@
-fileTitle,videoTitle,director,year,rating,marks
-null,null,null,0,0,0
-File Title,Video Title,Director,2021,8,5
 fileTitle,videoTitle,director,year,songTitle,songwriter,likes,views
 null,null,null,0,null,null,0,0
 File Title,Video Title,Director,2021,Song Title,Songwriter,50,1000
+fileTitle,videoTitle,director,year,rating,marks
+null,null,null,0,0,0
+File Title,Video Title,Director,2021,8,5
 fileTitle,videoTitle,director,year,podcastHost,podcastTopic,likes,views
 null,null,null,0,null,null,0,0
 fileTitle,videoTitle,director,2022,podcastHost,podcastHost,5,10
